+++ internal/crypto/Cargo.toml
@@ -27,7 +27,17 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
 base64 = "0.22.1"
+c2pa-status-tracker = { path = "../status-tracker", version = "0.1.0" }
+rasn = "0.18.0"
+rasn-ocsp = "0.18.0"
+rasn-pkix = "0.18.0"
 sha1 = "0.10.6"
+thiserror = "1.0.61"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+ureq = "2.4.0"
+url = "=2.5.2" # Can't advance to 2.5.3 until Unicode-3.0 license is reviewed.
+x509-parser = "0.16.0"
 
 [dependencies.chrono]
 version = "0.4.38"
