+++ internal/crypto/src/tests/mod.rs
@@ -24,3 +24,4 @@ wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
 mod base64;
 mod hash;
 mod internal;
+mod ocsp;
